+++ src/Order/Contracts/OrderFactory.php
@@ -19,11 +19,6 @@ interface OrderFactory
 {
     /**
      * Creates a new order from simple data arrays
-     *
-     * @param array $data
-     * @param array $items
-     *
-     * @return Order
      */
-    public function createFromDataArray(array $data, array $items): Order;
+    public function createFromDataArray(array $data, array $items, array|callable $hooks = null, array|callable $itemHooks = null): Order;
 }